Lesson03HandsOn
Created a simple Node project that set up and ran a switch case to see which grade the variable belonged in and printed it to the console. Also rewrote a given if...else statement into a ternary operation that printed a message to the HTML document depending on the value of the variable being checked int the operation.